requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pdf2image
Pillow
opencv-python-headless
requests
pandas
rapidfuzz
paddleocr
scikit-image
layoutparser
langdetect
sentence-transformers
torch
transformers>=4.30.0
scikit-learn
pytesseract
aiohttp
numpy
# LayoutLMv3 dependencies (optional, falls back to heuristic if not available)
# transformers>=4.30.0  # Already included above
# torch  # Already included above

